optee: refactor call queue code
Move the call queue handling logic off to a new file to allow for its use
in other translation units cleanly.

Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
