+++ drivers/tee/optee/call.c
@@ -15,91 +15,6 @@
 #include "optee_smc.h"
 #include "optee_bench.h"
 
-struct optee_call_waiter {
-	struct list_head list_node;
-	struct completion c;
-};
-
-static void optee_cq_wait_init(struct optee_call_queue *cq,
-			       struct optee_call_waiter *w)
-{
-	/*
-	 * We're preparing to make a call to secure world. In case we can't
-	 * allocate a thread in secure world we'll end up waiting in
-	 * optee_cq_wait_for_completion().
-	 *
-	 * Normally if there's no contention in secure world the call will
-	 * complete and we can cleanup directly with optee_cq_wait_final().
-	 */
-	mutex_lock(&cq->mutex);
-
-	/*
-	 * We add ourselves to the queue, but we don't wait. This
-	 * guarantees that we don't lose a completion if secure world
-	 * returns busy and another thread just exited and try to complete
-	 * someone.
-	 */
-	init_completion(&w->c);
-	list_add_tail(&w->list_node, &cq->waiters);
-
-	mutex_unlock(&cq->mutex);
-}
-
-static void optee_cq_wait_for_completion(struct optee_call_queue *cq,
-					 struct optee_call_waiter *w)
-{
-	wait_for_completion(&w->c);
-
-	mutex_lock(&cq->mutex);
-
-	/* Move to end of list to get out of the way for other waiters */
-	list_del(&w->list_node);
-	reinit_completion(&w->c);
-	list_add_tail(&w->list_node, &cq->waiters);
-
-	mutex_unlock(&cq->mutex);
-}
-
-static void optee_cq_complete_one(struct optee_call_queue *cq)
-{
-	struct optee_call_waiter *w;
-
-	list_for_each_entry(w, &cq->waiters, list_node) {
-		if (!completion_done(&w->c)) {
-			complete(&w->c);
-			break;
-		}
-	}
-}
-
-static void optee_cq_wait_final(struct optee_call_queue *cq,
-				struct optee_call_waiter *w)
-{
-	/*
-	 * We're done with the call to secure world. The thread in secure
-	 * world that was used for this call is now available for some
-	 * other task to use.
-	 */
-	mutex_lock(&cq->mutex);
-
-	/* Get out of the list */
-	list_del(&w->list_node);
-
-	/* Wake up one eventual waiting task */
-	optee_cq_complete_one(cq);
-
-	/*
-	 * If we're completed we've got a completion from another task that
-	 * was just done with its call to secure world. Since yet another
-	 * thread now is available in secure world wake up another eventual
-	 * waiting task.
-	 */
-	if (completion_done(&w->c))
-		optee_cq_complete_one(cq);
-
-	mutex_unlock(&cq->mutex);
-}
-
 /* Requires the filpstate mutex to be held */
 static struct optee_session *find_session(struct optee_context_data *ctxdata,
 					  u32 session_id)
