+++ drivers/tee/optee/optee_private.h
@@ -119,6 +119,11 @@ struct optee_rpc_param {
 	u32	a7;
 };
 
+struct optee_call_waiter {
+	struct list_head list_node;
+	struct completion c;
+};
+
 /* Holds context that is preserved during one STD call */
 struct optee_call_ctx {
 	/* information about pages list used in last allocation */
@@ -182,6 +187,14 @@ void optee_fill_pages_list(u64 *dst, struct page **pages, int num_pages,
 #define PTA_CMD_GET_DEVICES		0x0
 #define PTA_CMD_GET_DEVICES_SUPP	0x1
 int optee_enumerate_devices(u32 func);
+
+void optee_cq_wait_init(struct optee_call_queue *cq,
+			struct optee_call_waiter *w);
+void optee_cq_wait_for_completion(struct optee_call_queue *cq,
+				  struct optee_call_waiter *w);
+void optee_cq_complete_one(struct optee_call_queue *cq);
+void optee_cq_wait_final(struct optee_call_queue *cq,
+			 struct optee_call_waiter *w);
 
 /*
  * Small helpers
